+++ extensions/envato/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - View your purchases
 - Add placeholder to `Preference`
-- Fix type in title of a `Command`
+- Fix typo in title of a `Command`
 
 ## [Update Logo] - 2024-08-27